Convert GeoMapStore to a service
