+++ ui/main/src/app/services/geoMap/model/HighlightedCard.ts
@@ -1,0 +1,18 @@
+/* Copyright (c) 2024-2025, RTE (http://www.rte-france.com)
+ * See AUTHORS.txt
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ * This file is part of the OperatorFabric project.
+ */
+
+export class HighlightedCard {
+    public readonly cardId: string;
+    public readonly highlight: boolean;
+
+    constructor(cardId: string, highLight: boolean) {
+        this.cardId = cardId;
+        this.highlight = highLight;
+    }
+}
